genai-kids-backend/requirements.txt
openai
langchain
langgraph
python-dotenv
fastapi
uvicorn
langchain-huggingface
TTS